+++ src/components/common/header/nav.ts
@@ -1,4 +1,4 @@
-import { PATHS } from "@routes/paths";
+import { PATHS } from "@constants/paths";
 
 export const NAV_ITEMS = [
   { name: "홈", path: PATHS.HOME },
