+++ src/routes/index.tsx
@@ -1,8 +1,8 @@
 import { createBrowserRouter } from "react-router-dom";
 import NoteLayout from "src/layouts/noteLayout";
+import { PATHS } from "../constants/paths";
 import MainLayout from "../layouts/mainLayout";
 import { Archive, Flashcard, Home, Library, LoginPage, Mypage } from "../pages";
-import { PATHS } from "./paths";
 
 const router = createBrowserRouter([
   {
